Refactor shipping sections into partials
